+++ src/main/java/learn/poker/models/Game.java
@@ -6,14 +6,14 @@ import java.util.Objects;
 public class Game {
 
     private int gameId;
-    private int pot;
+    private double pot;
     private String winner;
 
     private Action lastAction;
     private Board board;
     private List<Player> players;
 
-    public Game(int gameId, int pot, String winner, Action lastAction, Board board, List<Player> players) {
+    public Game(int gameId, double pot, String winner, Action lastAction, Board board, List<Player> players) {
         this.gameId = gameId;
         this.pot = pot;
         this.winner = winner;
@@ -22,7 +22,7 @@ public class Game {
         this.players = players;
     }
 
-    public Game(int gameId, int pot, String winner, Board board, List<Player> players) {
+    public Game(int gameId, double pot, String winner, Board board, List<Player> players) {
         this.gameId = gameId;
         this.pot = pot;
         this.winner = winner;
@@ -30,7 +30,7 @@ public class Game {
         this.players = players;
     }
 
-    public Game(int pot, String winner, Board board, List<Player> players) {
+    public Game(double pot, String winner, Board board, List<Player> players) {
         this.pot = pot;
         this.winner = winner;
         this.board = board;
@@ -47,11 +47,11 @@ public class Game {
         this.gameId = gameId;
     }
 
-    public int getPot() {
+    public double getPot() {
         return pot;
     }
 
-    public void setPot(int pot) {
+    public void setPot(double pot) {
         this.pot = pot;
     }
